backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
python-jose==3.3.0
passlib[bcrypt]==1.7.4
motor==3.4.0
httpx==0.27.0
python-dotenv==1.0.1
pydantic-settings==2.2.1

